# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MonkeyCode API",
        "title": "MonkeyCode API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/security/scanning": {
            "get": {
                "description": "获取扫描结果",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Security Scanning"
                ],
                "summary": "获取扫描结果",
                "operationId": "security-scanning-list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "作者",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "下一页标识",
                        "name": "next_token",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "分页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目名称",
                        "name": "project_name",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "每页多少条记录",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.ListSecurityScanningResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/security/scanning/detail": {
            "get": {
                "description": "获取扫描风险详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Security Scanning"
                ],
                "summary": "获取扫描风险详情",
                "operationId": "security-scanning-detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "扫描任务id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.SecurityScanningRiskDetail"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/v1/completion/accept": {
            "post": {
                "description": "接受补全请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAIV1"
                ],
                "summary": "接受补全请求",
                "operationId": "accept-completion",
                "parameters": [
                    {
                        "description": "补全请求",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.AcceptCompletionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.Resp"
                        }
                    }
                }
            }
        },
...
        "/v1/report": {
            "post": {
                "description": "报告，支持多种操作：accept（接受补全）、suggest（建议）、reject（拒绝补全并记录用户输入）、file_written（文件写入）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAIV1"
                ],
                "summary": "报告",
                "operationId": "report",
                "parameters": [
                    {
                        "description": "报告请求",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.Resp"
                        }
                    }
                }
            }
        },
        "/v1/security/scanning": {
            "get": {
                "description": "分页逻辑只支持用 next_token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAIV1"
                ],
                "summary": "扫描任务列表",
                "operationId": "list-security-scanning",
                "parameters": [
                    {
                        "description": "扫描任务列表请求",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ListSecurityScanningReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.ListSecurityScanningBriefResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "创建扫描任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAIV1"
                ],
                "summary": "创建扫描任务",
                "operationId": "create-security-scanning",
                "parameters": [
                    {
                        "description": "创建扫描任务请求",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateSecurityScanningReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.Resp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "consts.ReportAction": {
            "type": "string",
            "enum": [
                "accept",
                "suggest",
                "file_written",
                "reject",
                "new_task",
                "feedback_task",
                "abort_task"
            ],
            "x-enum-varnames": [
                "ReportActionAccept",
                "ReportActionSuggest",
                "ReportActionFileWritten",
                "ReportActionReject",
                "ReportActionNewTask",
                "ReportActionFeedbackTask",
                "ReportActionAbortTask"
            ]
        },
        "consts.SecurityScanningLanguage": {
            "type": "string",
            "enum": [
                "C/C++",
                "Java",
                "Python",
                "JavaScript",
                "Go",
                "PHP",
                "C#",
                "Swift",
                "Ruby",
                "Rust",
                "HTML",
                "Objective-C/C++",
                "OCaml",
                "Kotlin",
                "Scala",
                "Solidity",
                "COBOL",
                "Shell",
                "SQL",
                "Fortran",
                "Dart",
                "Groovy",
                "Lua",
                "Secrets",
                "IaC"
            ],
            "x-enum-varnames": [
                "SecurityScanningLanguageCpp",
                "SecurityScanningLanguageJava",
                "SecurityScanningLanguagePython",
                "SecurityScanningLanguageJavaScript",
                "SecurityScanningLanguageGo",
                "SecurityScanningLanguagePHP",
                "SecurityScanningLanguageCS",
                "SecurityScanningLanguageSwift",
                "SecurityScanningLanguageRuby",
                "SecurityScanningLanguageRust",
                "SecurityScanningLanguageHTML",
                "SecurityScanningLanguageObjectiveC",
                "SecurityScanningLanguageOCaml",
                "SecurityScanningLanguageKotlin",
                "SecurityScanningLanguageScala",
                "SecurityScanningLanguageSolidity",
                "SecurityScanningLanguageCOBOL",
                "SecurityScanningLanguageShell",
                "SecurityScanningLanguageSQL",
                "SecurityScanningLanguageFortran",
                "SecurityScanningLanguageDart",
                "SecurityScanningLanguageGroovy",
                "SecurityScanningLanguageLua",
                "SecurityScanningLanguageSecrets",
                "SecurityScanningLanguageIaC"
            ]
        },
        "consts.SecurityScanningRiskLevel": {
            "type": "string",
            "enum": [
                "severe",
                "critical",
                "suggest"
            ],
            "x-enum-comments": {
                "SecurityScanningRiskLevelCritical": "高危",
                "SecurityScanningRiskLevelSevere": "严重",
                "SecurityScanningRiskLevelSuggest": "建议"
            },
            "x-enum-varnames": [
                "SecurityScanningRiskLevelSevere",
                "SecurityScanningRiskLevelCritical",
                "SecurityScanningRiskLevelSuggest"
            ]
        },
        "consts.SecurityScanningStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "success",
                "failed"
            ],
            "x-enum-varnames": [
                "SecurityScanningStatusPending",
                "SecurityScanningStatusRunning",
                "SecurityScanningStatusSuccess",
                "SecurityScanningStatusFailed"
            ]
        },
...
        "consts.UserStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive",
                "locked"
            ],
            "x-enum-varnames": [
                "UserStatusActive",
                "UserStatusInactive",
                "UserStatusLocked"
            ]
        },
        "domain.AcceptCompletionReq": {
            "type": "object",
            "properties": {
                "completion": {
                    "description": "补全内容",
                    "type": "string"
                },
                "id": {
                    "description": "记录ID",
                    "type": "string"
                }
            }
        },
...
        "domain.CreateSecurityScanningReq": {
            "type": "object",
            "properties": {
                "language": {
                    "description": "扫描语言",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.SecurityScanningLanguage"
                        }
                    ]
                },
                "workspace": {
                    "description": "项目目录",
                    "type": "string"
                }
            }
        },
...
        "domain.ListSecurityScanningBriefResp": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.SecurityScanningBrief"
                    }
                },
                "next_token": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "domain.ListSecurityScanningReq": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "作者",
                    "type": "string"
                },
                "next_token": {
                    "description": "下一页标识",
                    "type": "string"
                },
                "page": {
                    "description": "分页",
                    "type": "integer",
                    "default": 1,
                    "minimum": 1
                },
                "project_name": {
                    "description": "项目名称",
                    "type": "string"
                },
                "size": {
                    "description": "每页多少条记录",
                    "type": "integer",
                    "default": 10,
                    "minimum": 1
                }
            }
        },
        "domain.ListSecurityScanningResp": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.SecurityScanningResult"
                    }
                },
                "next_token": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "domain.ReportReq": {
            "type": "object",
            "properties": {
                "action": {
                    "$ref": "#/definitions/consts.ReportAction"
                },
                "content": {
                    "description": "内容",
                    "type": "string"
                },
                "cursor_position": {
                    "description": "光标位置（用于reject action）",
                    "type": "object",
                    "additionalProperties": {}
                },
                "id": {
                    "description": "task_id or resp_id",
                    "type": "string"
                },
                "mode": {
                    "description": "模式",
                    "type": "string"
                },
                "source_code": {
                    "description": "当前文件的原文（用于reject action）",
                    "type": "string"
                },
                "tool": {
                    "description": "工具",
                    "type": "string"
                },
                "user_input": {
                    "description": "用户输入的新文本（用于reject action）",
                    "type": "string"
                }
            }
        },
        "domain.SecurityScanningBrief": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "report_url": {
                    "description": "报告url",
                    "type": "string"
                },
                "status": {
                    "description": "扫描状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.SecurityScanningStatus"
                        }
                    ]
                },
                "workspace": {
                    "description": "项目目录",
                    "type": "string"
                }
            }
        },
        "domain.SecurityScanningResult": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "扫描开始时间",
                    "type": "integer"
                },
                "id": {
                    "description": "扫描任务id",
                    "type": "string"
                },
                "name": {
                    "description": "扫描任务",
                    "type": "string"
                },
                "project_name": {
                    "description": "项目名称",
                    "type": "string"
                },
                "risk": {
                    "description": "风险结果",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.SecurityScanningRiskResult"
                        }
                    ]
                },
                "status": {
                    "description": "扫描状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.SecurityScanningStatus"
                        }
                    ]
                },
                "user": {
                    "description": "用户",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.User"
                        }
                    ]
                }
            }
        },
        "domain.SecurityScanningRiskDetail": {
            "type": "object",
            "properties": {
                "desc": {
                    "description": "风险描述",
                    "type": "string"
                },
                "filename": {
                    "description": "风险文件名",
                    "type": "string"
                },
                "id": {
                    "description": "风险id",
                    "type": "string"
                },
                "level": {
                    "description": "风险等级",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.SecurityScanningRiskLevel"
                        }
                    ]
                }
            }
        },
        "domain.SecurityScanningRiskResult": {
            "type": "object",
            "properties": {
                "critical_count": {
                    "description": "高危数",
                    "type": "integer"
                },
                "severe_count": {
                    "description": "严重数",
                    "type": "integer"
                },
                "suggest_count": {
                    "description": "建议数",
                    "type": "integer"
                }
            }
        },
...
        "domain.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "description": "头像URL",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "用户ID",
                    "type": "string"
                },
                "is_deleted": {
                    "description": "是否删除",
                    "type": "boolean"
                },
                "last_active_at": {
                    "description": "最后活跃时间",
                    "type": "integer"
                },
                "status": {
                    "description": "用户状态 active: 正常 locked: 锁定 inactive: 禁用",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.UserStatus"
                        }
                    ]
                },
                "two_step_auth": {
                    "description": "是否开启两步验证",
                    "type": "boolean"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "web.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}